refactor(interpolate): keep only the tile forward kernel
