+++ crates/cubek-interpolate/src/definition/mod.rs
@@ -10,7 +10,5 @@ pub use modes::*;
 pub use problem::*;
 pub use transform::*;
 
-pub use crate::multi_level::{InterpolatePrecision, TileSize};
-pub use crate::multi_level::tile_size::tile_absolute_coords;
 mod base;
 pub use base::*;
