+++ crates/cubek-interpolate/src/kernel/forward/coordinate.rs
