+++ crates/cubek-interpolate/src/kernel/forward/filter.rs
