+++ crates/cubek-interpolate/src/kernel/forward/geometry.rs
