+++ crates/cubek-interpolate/src/kernel/forward/space.rs
