+++ crates/cubek-interpolate/src/kernel/mod.rs
@@ -1,0 +1,6 @@
+pub mod backward;
+pub mod forward;
+
+pub use forward::TileConfig;
+
+pub(crate) use backward::interpolate_nearest_backward_launch;
