--- crates/cubek-interpolate/src/launch.rs
@@ -1,3 +1,0 @@
-pub(crate) use crate::multi_level::launch::{
-    interpolate_launch, interpolate_nearest_backward_launch,
-};
