+++ crates/cubek-interpolate/src/lib.rs
@@ -1,16 +1,12 @@
 pub mod definition;
-pub mod multi_level;
 #[cfg(any(feature = "cpu-reference", feature = "benchmarks"))]
 pub mod eval;
-mod launch;
-pub mod strategy;
-pub mod tiled;
+pub mod kernel;
 pub mod tune_key;
 
 use crate::{
     definition::{InterpolateError, InterpolateMode, InterpolateOptions},
-    launch::{interpolate_launch, interpolate_nearest_backward_launch},
-    strategy::Strategy,
+    kernel::{TileConfig, forward::interpolate_launch, interpolate_nearest_backward_launch},
 };
 use core::result::Result;
 use cubecl::{Runtime, client::ComputeClient, prelude::TensorBinding, prelude::*};
@@ -25,14 +21,13 @@ pub fn interpolate<R: Runtime>(
     input: TensorBinding<R>,
     output: TensorBinding<R>,
     options: InterpolateOptions,
-    strategy: Strategy,
+    config: TileConfig,
     dtype: ElemType,
 ) -> Result<(), InterpolateError> {
-    validate_strategy(client, &strategy)?;
     validate_rank(input.shape.len(), output.shape.len())?;
     validate_nhwc_consistency(&input.shape, &output.shape)?;
 
-    interpolate_launch(client, input, output, options, strategy, dtype)
+    interpolate_launch(client, input, output, options, dtype, config)
 }
 
 /// Backward interpolate operation
@@ -68,25 +63,6 @@ pub fn interpolate_backward<R: Runtime>(
             "{:?} interpolation backward is not supported by JIT backend",
             options.mode
         ))),
-    }
-}
-
-/// Checks if the strategy is valid for the current client
-fn validate_strategy<R: Runtime>(
-    client: &ComputeClient<R>,
-    strategy: &Strategy,
-) -> Result<(), InterpolateError> {
-    // If the client is not running on a CPU, we don't need to validate the strategy
-    if client.properties().hardware.num_cpu_cores.is_none() {
-        return Ok(());
-    }
-
-    // If the client is running on a CPU, we need to validate the strategy
-    match strategy {
-        Strategy::GlobalMemoryStrategy(_) => Ok(()),
-        Strategy::SharedMemoryStrategy(_) => Err(InterpolateError::UnsupportedMode(
-            "interpolation shared memory strategy is not supported on CPU".to_string(),
-        )),
     }
 }
