--- crates/cubek-interpolate/src/multi_level/mod.rs
@@ -1,10 +1,0 @@
-pub mod components;
-pub mod launch;
-pub mod precision;
-pub mod routines;
-pub mod settings;
-pub mod strategy;
-pub mod tile_size;
-
-pub use precision::{InterpolatePrecision, accumulator_dtype};
-pub use tile_size::TileSize;
