--- crates/cubek-interpolate/src/multi_level/strategy.rs
@@ -1,7 +1,0 @@
-use crate::multi_level::routines::{BlueprintStrategy, GlobalMemoryRoutine, SharedMemoryRoutine};
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum InterpolateStrategy {
-    GlobalMemoryStrategy(BlueprintStrategy<GlobalMemoryRoutine>),
-    SharedMemoryStrategy(BlueprintStrategy<SharedMemoryRoutine>),
-}
