--- crates/cubek-interpolate/src/strategy.rs
@@ -1,1 +1,0 @@
-pub use crate::multi_level::strategy::InterpolateStrategy as Strategy;
